openstack: switch the default json output to dictionnary
The python-openstackclient output may either be a list of Value/Field or
a dictionnary. Use the dictionnary instead of the Value/Field form
because it is easier to deal with. The former was chosen because of a
bug that made it impossible to control the output. The bug still exists
but is believed to be rare enough to ignore. Should it happen, a message
recommending a corrective action and a comment on:

https://bugs.launchpad.net/python-openstackclient/+bug/1510546

is displayed.

Signed-off-by: Loic Dachary <loic@dachary.org>